changed the endpoint to /schemas/fieldSpec
